Moved VCTHDFactory to .Internal namespace
